+++ docs/src/IconButton.vue
@@ -1,8 +1,9 @@
 <script lang="ts" setup>
-import { computed, nextTick, ref } from 'vue'
-import Icon, { IconActionDeleteMedium } from '@cypress-design/vue-icon'
+import { computed, nextTick, ref, watch } from 'vue'
+import { IconActionDeleteMedium } from '@cypress-design/vue-icon'
 import _ from 'lodash'
 import CopyButton from './CopyButton.vue'
+import IconSized from './IconSized.vue'
 
 const { upperFirst, camelCase } = _
 
@@ -17,8 +18,6 @@ const localFocused = ref(false)
 
 const focused = computed(() => props.focused || localFocused.value)
 
-const IconAny = Icon as any
-
 const placeholderStyle = ref({
   width: '1px',
   height: '1px',
@@ -30,28 +29,56 @@ const buttonStyle = ref({
   top: '0',
   height: '0',
   transform: 'none',
+  transformOrigin: 'left center',
+  transition: 'none',
+})
+
+watch(localFocused, (value) => {
+  if (!value) {
+    buttonStyle.value = {
+      ...buttonStyle.value,
+      transform: 'none',
+      transition: 'none',
+    }
+  }
 })
 
 function focus() {
+  const buttonWidth = $button.value?.offsetWidth ?? 0
+  const buttonLeft = $button.value?.offsetLeft ?? 0
+
   placeholderStyle.value = {
     top: `${$button.value?.offsetTop}px`,
     height: `${$button.value?.offsetHeight}px`,
-    width: `${$button.value?.offsetWidth}px`,
-    left: `${$button.value?.offsetLeft}px`,
+    width: `${buttonWidth}px`,
+    left: `${buttonLeft}px`,
   }
 
+  const completeWidth = $button.value?.offsetParent?.clientWidth ?? 1
+
   buttonStyle.value = {
+    ...buttonStyle.value,
+    transform: `translateX(${buttonLeft - 36}px) scaleX(${
+      buttonWidth / (completeWidth - 74)
+    })`,
     top: `${($button.value?.offsetTop ?? 0) + 4}px`,
     height: `${($button.value?.offsetHeight ?? 0) - 8}px`,
-    transform: 'scaleX(1)',
   }
 
   nextTick(() => {
     localFocused.value = true
+    setTimeout(() => {
+      buttonStyle.value = {
+        ...buttonStyle.value,
+        transition: 'transform 0.15s linear',
+        transform: 'none',
+      }
+    }, 50)
   })
 }
 
 const $button = ref<HTMLDivElement>()
+// <tw-include class="grid-cols-1 grid-cols-2 grid-cols-3 grid-cols-4"/>
 </script>
 
 <template>
@@ -67,24 +94,26 @@ const $button = ref<HTMLDivElement>()
   />
   <button
     ref="$button"
-    class="gap-x-[16px] flex flex-wrap overflow-hidden bg-indigo-50 dark:bg-gray-800"
+    class="gap-x-[16px] flex flex-wrap items-center overflow-hidden bg-indigo-50 dark:bg-gray-800 min-h-[72px] rounded"
     :class="{
-      'mx-[16px] px-[8px] pb-[4px] rounded md:flex-nowrap justify-end md:justify-start':
+      'mx-[16px] px-[8px] pb-[4px] md:flex-nowrap justify-end md:justify-start !cursor-default':
         focused,
-      'rounded py-[8px] justify-center': !focused,
-      'absolute left-0 right-0 md:left-[28px] md:right-[28px] z-20 w-auto items-center min-h-[120px] md:min-h-0 transition-transform':
+      'py-[8px] justify-center hover:bg-indigo-100 dark:hover:bg-gray-700 transition-colors':
+        !focused,
+      'absolute left-0 right-0 md:left-[28px] md:right-[28px] z-20 w-auto items-center min-h-[120px] md:min-h-0':
         localFocused,
       'w-[calc(100%-32px)] lg:w-[700px] mx-auto items-end':
         !localFocused && focused,
     }"
-    @click="focus()"
     :style="localFocused ? buttonStyle : undefined"
+    @click="!focused ? focus() : undefined"
   >
     <button
       v-if="localFocused"
       class="absolute top-[4px] right-[4px] rounded-full border-2 border-solid border-transparent hover:border-gray-500 dark:hover:border-gray-500"
       @click.stop="localFocused = false"
     >
+      <span class="sr-only">Close</span>
       <IconActionDeleteMedium />
     </button>
     <p
@@ -95,62 +124,32 @@ const $button = ref<HTMLDivElement>()
       }"
     >
       <span class="flex items-center md:justify-end mb-[8px] gap-x-[8px] group"
-        ><CopyButton :text="iconName" /><code>{{ iconName }}</code></span
+        ><CopyButton :text="iconName" /><code class="!m-0">{{
+          iconName
+        }}</code></span
       >
       <span class="flex items-center md:justify-end gap-x-[8px] group"
         ><CopyButton
           :text="`&lt;Icon${upperFirst(camelCase(iconName))}/&gt;`"
-        /><code>&lt;Icon{{ upperFirst(camelCase(iconName)) }} /&gt;</code></span
+        /><code class="!m-0"
+          >&lt;Icon{{ upperFirst(camelCase(iconName)) }} /&gt;</code
+        ></span
       >
     </p>
+
     <div
-      v-for="size in meta.availableSizes"
-      :key="size"
-      class="flex gap-[8px] items-end"
+      class="flex-grow grid gap-[16px] transition-all duration-1000"
+      :class="`grid-cols-${Math.min(meta.availableSizes.length, 4)}`"
     >
-      <div
-        class="py-[4px] min-w-[32px] flex flex-col items-center gap-x-[16px] gap-y-[4px] justify-end"
-        :class="{
-          'pl-[4px] border-l border-gray-300': focused,
-          'px-[8px] md:px-[4px] md:w-[164px]': !focused,
-        }"
-      >
-        <IconAny :name="iconName" :size="size" />
-        <p class="text-gray-500 text-[12px] mt-1">
-          <span v-if="!focused">{{
-            iconName.slice(groupName.length + 1)
-          }}</span>
-          {{ size }}
-        </p>
-      </div>
-      <div
-        v-if="focused"
-        :key="`${iconName}_${size}`"
-        class="text-center text-teal-500"
-      >
-        <div v-if="meta.hasStrokeColor && meta.hasStrokeColor.includes(size)">
-          s
-        </div>
-        <div v-if="meta.hasFillColor && meta.hasFillColor.includes(size)">
-          f
-        </div>
-        <div
-          v-if="
-            meta.hasSecondaryStrokeColor &&
-            meta.hasSecondaryStrokeColor.includes(size)
-          "
-        >
-          s+
-        </div>
-        <div
-          v-if="
-            meta.hasSecondaryFillColor &&
-            meta.hasSecondaryFillColor.includes(size)
-          "
-        >
-          f+
-        </div>
-      </div>
+      <IconSized
+        v-for="size in meta.availableSizes"
+        :key="size"
+        :focused="focused"
+        :iconName="iconName"
+        :groupName="groupName"
+        :size="size"
+        :meta="meta"
+      />
     </div>
   </button>
 </template>