+++ test/webpack-react-app/webpack.config.js
@@ -1,4 +1,5 @@
 const path = require('path')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
 const { CyCSSWebpackPlugin } = require('@cypress-design/css')
 
 const config = {
@@ -26,9 +27,7 @@ const config = {
     ],
   },
   devServer: {
-    static: {
-      directory: './public',
-    },
+    port: 5173,
   },
   resolve: {
     extensions: ['.tsx', '.ts', '.js', '.css'],
@@ -44,6 +43,13 @@ const config = {
           '../../components/*/react/dist/*.@(js|css)',
         ],
       },
+    }),
+    new HtmlWebpackPlugin({
+      templateContent: ({ htmlWebpackPlugin }) =>
+        '<!DOCTYPE html><html><head><meta charset="utf-8"><title>' +
+        htmlWebpackPlugin.options.title +
+        '</title></head><body><div id="app"></div></body></html>',
+      filename: 'index.html',
     }),
   ],
 }