+++ components/StatusIcon/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cypress-design/react-statusicon
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [[`bc28be1`](https://github.com/cypress-io/cypress-design/commit/bc28be1760cf7183bbfb6d66f5362ad10ce22d34)]:
+  - @cypress-design/react-icon@0.21.0
+
 ## 0.4.1
 
 ### Patch Changes