+++ test/vue-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cypress-design/test-vue-app
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`bc28be1`](https://github.com/cypress-io/cypress-design/commit/bc28be1760cf7183bbfb6d66f5362ad10ce22d34)]:
+  - @cypress-design/vue-icon@0.22.0
+
 ## 0.1.1
 
 ### Patch Changes